Save failed script results in WardenError
Staging is run by the DEA in a container via the 'run_script' facility.
Because of how loggregator reads staging messages, failures during
staging prevent any diagnostic output from being presented to the user.
In order to capture some of this diagnostic info without radically
restructuring code, the Container::WardenError has been extended to
capture and hold the failed script results for the caller to access.